Remove thematic break from block refs, move block ref logic into plugin
